eval/geneval_requirements.txt
addict==2.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
async-timeout==5.0.1
attrs==25.3.0
braceexpand==0.1.7
brotlipy==0.7.0
certifi==2023.11.17
cffi==1.15.1
charset-normalizer==2.0.4
clip-benchmark==1.6.1
colorama==0.4.6
contourpy==1.3.0
cryptography==39.0.1
cycler==0.12.1
datasets==2.11.0
diffusers==0.24.0
dill==0.3.6
filelock==3.18.0
fonttools==4.58.4
frozenlist==1.7.0
fsspec==2025.5.0
ftfy==6.3.1
hf-xet==1.1.4
huggingface-hub==0.33.0
idna==3.4
importlib-metadata==8.7.0
importlib-resources==6.5.2
joblib==1.5.1
kiwisolver==1.4.7
markdown-it-py==3.0.0
matplotlib==3.9.4
mdurl==0.1.2
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
mmcv-full==1.6.2
mmengine==0.10.7
multidict==6.5.0
multiprocess==0.70.14
numpy==1.23.1
open-clip-torch==2.20.0
opencv-python==4.11.0.86
packaging==25.0
pandas==2.3.0
Pillow==9.4.0
platformdirs==4.3.8
propcache==0.3.2
protobuf==3.20.3
pyarrow==20.0.0
pycocoevalcap==1.2
pycocotools==2.0.10
pycparser==2.21
pygments==2.19.1
pyOpenSSL==23.0.0
pyparsing==3.2.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0
regex==2022.4.24
requests==2.29.0
responses==0.18.0
rich==14.0.0
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.13.1
sentencepiece==0.2.0
six==1.16.0
termcolor==3.1.0
terminaltables==3.1.10
threadpoolctl==3.6.0
timm==1.0.15
tokenizers==0.21.1
tomli==2.2.1
torch==1.12.1
torchvision==0.13.1
tqdm==4.67.1
transformers==4.52.4
typing-extensions==4.5.0
tzdata==2025.2
urllib3==2.0.2
wcwidth==0.2.13
webdataset==0.2.100
xxhash==3.5.0
yapf==0.43.0
yarl==1.20.1
zipp==3.22.0